Even more 64-bit fixes.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@3540 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Frameworks/EtoileSerialize/Headers/ETDeserializer.h
@@ -11,7 +11,7 @@
 typedef struct 
 {
 	void * startOffset;
-	int index;
+	unsigned long index;
 	char type;
 	int size;
 } ETDeserializerState;
